Convert known issue tests to regression tests.
